+++ contracts/interfaces/IZone.sol
@@ -7,21 +7,21 @@ interface IZone {
     error OnlyFactory();
 
     function tokenId() external view returns (uint256);
-    function metadata() external view returns (uint24, uint24, uint256);
+    function metadata() external view returns (uint64, uint64, uint256);
     function factory() external view returns (address);
     function shareToken(uint256) external view returns (address);
     function marketplace() external view returns (address);
-    function initialize(string memory, string memory, uint24, uint24, address, IShareTokenRegistry) external;
+    function initialize(string memory, string memory, uint64, uint64, address, IShareTokenRegistry) external;
     function mint(
         address,
         uint256,
         string memory,
         address,
-        uint24[4] memory,
+        uint64[4] memory,
         uint8
     ) external returns (uint256, address);
     function burn(uint256) external returns (address);
-    function trades(uint256) external view returns (uint24);
+    function trades(uint256) external view returns (uint64);
     function tradesLength() external view returns (uint256);
     function updateTrades(bool) external;
     function setMarketplace(address) external;
@@ -29,5 +29,5 @@ interface IZone {
     function exists(uint256) external view returns (bool);
 
     event Mint(uint256 indexed tokenId, address indexed shareToken, string metadataURI);
-    event Initialize(string name, string symbol, uint24 lat, uint24 lng);
+    event Initialize(string name, string symbol, uint64 lat, uint64 lng);
 }
